+++ R/LUCS_KDD_CBA.R
@@ -47,8 +47,7 @@
 #'
 #' Yin, Xiaoxin and Jiawei Han. CPAR: Classification based on Predictive
 #' Association Rules, SDM, 2003.
-#' c("\\Sexpr[results=rd,stage=build]{tools:::Rd_expr_doi(\"#1\")}",
-#' "10.1137/1.9781611972733.40")\Sexpr{tools:::Rd_expr_doi("10.1137/1.9781611972733.40")}
+#' \doi{10.1137/1.9781611972733.40}
 #'
 #' Frans Coenen et al. The LUCS-KDD Software Library,
 #' \url{https://cgi.csc.liv.ac.uk/~frans/KDD/Software/}
@@ -204,17 +203,15 @@ FOIL2 <- function(formula, data, best_k = 5, disc.method = "mdlp", verbose = FAL
 
   rules <- .LUCS_KDD(formula, trans, method = "FOIL", parameter = "", verbose = verbose)
 
-  structure(list(
+  CBA_ruleset(
     formula = formula,
-    discretization = attr(trans, "disc_info"),
     rules = rules,
     default = majorityClass(formula, trans),
     method = "weighted",
     weights = "laplace",
     best_k = best_k,
-    description = paste0("FOIL-based classifier (Yin and Han, 2003 - LUCS-KDD implementation).")
-  ),
-    class = "CBA"
+    discretization = attr(trans, "disc_info"),
+    description = "FOIL-based classifier (Yin and Han, 2003 - LUCS-KDD implementation)."
   )
 }