+++ R/RCAR.R
@@ -1,19 +1,19 @@
 #' Regularized Class Association Rules for Multi-class Problems (RCAR+)
-#' 
+#'
 #' Build a classifier based on association rules mined for an input dataset and
 #' weighted with LASSO regularized logistic regression following RCAR (Azmi, et
 #' al., 2019). RCAR+ extends RCAR from a binary classifier to a multi-class
 #' classifier and can use support-balanced CARs.
-#' 
+#'
 #' RCAR+ extends RCAR from a binary classifier to a multi-class classifier
 #' using regularized multinomial logistic regression via \pkg{glmnet}.
-#' 
+#'
 #' If lambda is not specified (\code{NULL}) then cross-validation with the
 #' largest value of lambda such that error is within 1 standard error of the
 #' minimum is used to determine the best value (see \code{\link{cv.glmnet}}).
-#' 
+#'
 #' See \code{\link{cv.glmnet}} for performing cross-validation in parallel.
-#' 
+#'
 #' @aliases RCAR rcar
 #' @param formula A symbolic description of the model to be fitted. Has to be
 #' of form \code{class ~ .} or \code{class ~ predictor1 + predictor2}.
@@ -39,7 +39,7 @@
 #' @return Returns an object of class \code{CBA} representing the trained
 #' classifier with the additional field \code{model} containing a list with the
 #' following elements:
-#' 
+#'
 #' \item{all_rules}{all rules used to build the classifier, including the rules
 #' with a weight of zero.} \item{reg_model}{them multinomial logistic
 #' regression model as an object of class \code{\link{glmnet}}.}
@@ -52,28 +52,28 @@
 #' regularized class association rules algorithm for classification in a
 #' categorical data space. \emph{Information Sciences,} Volume 483, May 2019.
 #' Pages 313-331.
-#' 
+#'
 #' Azmi's implementation on GitHub: \url{https://github.com/azemi/RCAR}.
 #' @examples
-#' 
+#'
 #' data("iris")
-#' 
+#'
 #' classifier <- RCAR(Species~., iris)
 #' classifier
-#' 
+#'
 #' # inspect the rule base sorted by the larges class weight
 #' inspect(sort(rules(classifier), by = "weight"))
-#' 
+#'
 #' # make predictions for the first few instances of iris
 #' predict(classifier, head(iris))
-#' 
+#'
 #' # inspecting the regression model and the cross-validation results to determine lambda
 #' str(classifier$model$reg_model)
 #' plot(classifier$model$cv)
-#' 
+#'
 #' # show progress report and use 5 instead of the default 10 cross-validation folds.
 #' classifier <- RCAR(Species~., iris, cv.glmnet.args = list(nfolds = 5), verbose = TRUE)
-#' 
+#'
 RCAR <- function(formula, data,
   lambda = NULL, alpha = 1, glmnet.args = NULL, cv.glmnet.args = NULL,
   parameter = NULL, control = NULL, balanceSupport = FALSE,
@@ -119,24 +119,23 @@ RCAR <- function(formula, data,
   quality(cars)$oddsratio <- exp(quality(cars)$weight)
   rulebase <- cars[!remove]
   weights <- weights[!remove,]
-  biases <- model$a0
+  bias <- model$a0
 
   if(verbose) cat("* CARs left:", length(rulebase), "\n")
 
 
-  structure(list(
+  CBA_ruleset(
     formula = formula,
-    discretization = attr(trans, "disc_info"),
     rules = rulebase,
-    default = NA,
     weights = weights,
-    biases = biases,
+    bias = bias,
     method='logit',
     model = list(
       all_rules = cars,
       reg_model = model,
       cv = cv
     ),
-    description = paste("RCAR+ based on RCAR (Azmi et al., 2019)")
-  ), class = 'CBA')
+    discretization = attr(trans, "disc_info"),
+    description = "RCAR+ based on RCAR (Azmi et al., 2019)"
+  )
 }